Refactor: Rename templates.py to templating.py
Renames the file and updates all corresponding import statements to align with project conventions. Fixes #350
