+++ src/air/__init__.py
@@ -132,4 +132,4 @@ from .tags import (
     Video as Video,
     Wbr as Wbr,
 )
-from .templates import JinjaRenderer as JinjaRenderer
+from .templating import JinjaRenderer as JinjaRenderer
